requirements.txt
# Core dependencies
requests>=2.31.0
PyGithub>=2.1.1
pyparsing>=3.1.1
pydantic>=2.5.0
click>=8.1.7
rich>=13.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Additional utilities
tqdm>=4.66.1
